+++ pdfalyzer/output/styles/node_colors.py
@@ -2,28 +2,39 @@
 Configurations and methods to help with consistent styling of PDF trees, nodes, etc.
 """
 import re
+from collections import namedtuple
 from numbers import Number
+from typing import Any
 
-from PyPDF2.generic import ByteStringObject, IndirectObject
+from PyPDF2.generic import (ArrayObject, ByteStringObject, EncodedStreamObject, IndirectObject,
+     StreamObject, TextStringObject)
 from yaralyzer.output.rich_console import YARALYZER_THEME_DICT
 
 from pdfalyzer.helpers.pdf_object_helper import pypdf_class_name
 from pdfalyzer.output.styles.rich_theme import PDF_ARRAY
 from pdfalyzer.util import adobe_strings
 
+ClassStyle = namedtuple('ClassStyle', ['klass', 'style'])
+
 DEFAULT_LABEL_STYLE = 'yellow'
 FONT_OBJ_BLUE = 'deep_sky_blue4 bold'
 PDF_NON_TREE_REF = 'color(243)'
 
 # Subclasses of the key type will be styled with the value string
-NODE_TYPE_STYLES = {
-    Number: 'bright_cyan bold',
-    dict: 'color(64)',
-    list: 'color(143)',
-    str: 'bright_white bold',
-    IndirectObject: 'color(157)',
-    ByteStringObject: 'bytes',
-}
+NODE_TYPE_STYLES = [
+    ClassStyle(Number, 'cyan bold'),
+    ClassStyle(IndirectObject, 'color(225)'),
+    ClassStyle(ByteStringObject, 'bytes'),
+    ClassStyle(EncodedStreamObject, YARALYZER_THEME_DICT['bytes']),
+    ClassStyle(StreamObject, YARALYZER_THEME_DICT['bytes.title']),
+    ClassStyle(TextStringObject, YARALYZER_THEME_DICT['grey.light']),
+    ClassStyle(ArrayObject, PDF_ARRAY),
+    ClassStyle(dict, 'color(64)'),
+    ClassStyle(list, 'color(143)'),
+    ClassStyle(str, 'bright_white bold'),
+    # Default
+    ClassStyle(object, 'bright_yellow'),
+]
 
 LABEL_STYLES = [
     [re.compile('JavaScript|JS|OpenAction', re.I | re.M), 'blink bold red'],
@@ -31,6 +42,7 @@ LABEL_STYLES = [
     [re.compile(f'^{adobe_strings.FONT_FILE}'),           'steel_blue1'],
     [re.compile(f'^{adobe_strings.FONT}'),                FONT_OBJ_BLUE],
     [re.compile(f'^{adobe_strings.TO_UNICODE}'),          'grey30'],
+    [re.compile(f'^{adobe_strings.ENCODING}'),            YARALYZER_THEME_DICT['encoding.header']],
     [re.compile(f'^{adobe_strings.WIDTHS}'),              'color(67)'],
     [re.compile(f'^{adobe_strings.W}'),                   'color(67)'],
     [re.compile(f'^{adobe_strings.RESOURCES}'),           'magenta'],
@@ -57,38 +69,25 @@ LABEL_STYLES += [
 ]
 
 
-def get_type_style(klass) -> str:
+def get_class_style(obj: Any) -> str:
     """Style for various types of data (e.g. DictionaryObject)"""
-    return next((NODE_TYPE_STYLES[t] for t in NODE_TYPE_STYLES.keys() if issubclass(klass, t)), '')
+    return next((cs.style for cs in NODE_TYPE_STYLES if isinstance(obj, cs.klass)), '')
 
 
-def get_type_string_style(klass) -> str:
-    """Dim version of get_type_style() for non primitives, white for primitives"""
-    if issubclass(klass, (str, Number)):
-        return 'white'
+def get_class_style_dim(obj: Any) -> str:
+    """Dim version of get_class_style() for non primitives, white for primitives"""
+    if isinstance(obj, str):
+        return 'color(244)'
+    elif isinstance(obj, Number):
+        return 'cyan dim'
     else:
-        return f"{get_type_style(klass)} dim"
+        return f"{get_class_style(obj)} dim"
+
+
+def get_class_style_italic(obj: Any) -> str:
+    return f"{get_class_style(obj)} italic"
 
 
 def get_label_style(label: str) -> str:
     """Lookup a style based on the node's label string (either its type or first address)."""
     return next((ls[1] for ls in LABEL_STYLES if ls[0].search(label)), DEFAULT_LABEL_STYLE)
-
-
-def get_node_type_style(obj) -> str:
-    klass_string = pypdf_class_name(obj)
-
-    if 'Dictionary' in klass_string:
-        style = NODE_TYPE_STYLES[dict]
-    elif 'EncodedStream' in klass_string:
-        style = YARALYZER_THEME_DICT['bytes']
-    elif 'Stream' in klass_string:
-        style = YARALYZER_THEME_DICT['bytes.title']
-    elif 'Text' in klass_string:
-        style = YARALYZER_THEME_DICT['grey.light']
-    elif 'Array' in klass_string:
-        style = PDF_ARRAY
-    else:
-        style = 'bright_yellow'
-
-    return f"{style} italic"